originalPS/requirements.txt
scipy
numpy
pyyaml
Pillow==6.2.2
torchvision
torch

